+++ ChangeLog.md
@@ -18,3 +18,10 @@
 
 ## Unreleased changes
 - hoistOp
+- Renamed `Control.Category.FreeEff` module as `Control.Category.FreeEffect`
+  and renamed top level terms:
+    - `EffCategory` type class to `EffectCategory`
+    - `FreeEffCat` to `EffCat`
+    - `FreeEffCat` constructor as `Effect` and `lift` as `effect`
+    - `liftCat` to `liftEffect`
+    - `foldNatLift` to `foldNatEffCat`